fix: fixed bug with filter
